Resolv's methods return objects, not strings.
Resolv's `getaddress()` and `getname()` both return objects now that
have to be converted to strings before using them.

I suspect that Resolv changed at some-point in the past.  While the
new changes to the specs won't work with older rubies, the code
should work with older versions of jruby.

Fixes #11